Image-Captioning/requirements.txt
opencv-python>=3.2.0.6
jupyter>=1.0.0
matplotlib>=2.1.1
pandas>=0.22.0
numpy>=1.12.1
pillow>=5.0.0
scipy>=1.0.0
nltk>=3.2.2
tqdm>=4.19.4
seaborn>=0.8.1
torch>=0.4.0
torchvision>=0.2.0
